+++ README.md
@@ -34,7 +34,4 @@ y cómo quieres visualizarlos y sea capaz de sacarte el gráfico.
 
 [3. Funcionamiento correcto del despliegue en el PaaS (no sólo el status). Es decir, no se puede devolver ningún status 500.](docs/funcionamiento_correcto.md)
 
-[4. Buenas prácticas en el diseño del API, incluyendo su correspondencia correcta con diferentes HUs.]()
-
-[5. Uso correcto de bases de datos y logs dentro del PaaS, incluyendo su justificación y pruebas de prestaciones, así 
-como avance general y grado de terminación de la aplicación]()
+[4. Buenas prácticas en el diseño del API, incluyendo su correspondencia correcta con diferentes HUs.](docs/practicas_hus_api.md)
