+++ iv.yaml
@@ -3,4 +3,46 @@ test: tests/unit/BuildScatterPlot.test.js
 entidad: src/modules/Plots/interfaces/BuildScatterPlotInterface.js
 taskfile: package.json
 URL: https://us-central1-viz-your-data.cloudfunctions.net/api/plotTypes
+PaaS: https://viz-your-data-prod.herokuapp.com/
 make: grunt
+recurso:
+  nombre: /plotTypes/scatter
+  metodo: POST
+  payload: {
+    "x": "GrLivArea",
+    "y": "SalePrice",
+    "data": [
+      {
+        "GrLivArea": 1710,
+        "SalePrice": 208500
+      },
+      {
+        "GrLivArea": 1262,
+        "SalePrice": 181500
+      },
+      {
+        "GrLivArea": 1786,
+        "SalePrice": 223500
+      },
+      {
+        "GrLivArea": 1717,
+        "SalePrice": 140000
+      },
+      {
+        "GrLivArea": 2198,
+        "SalePrice": 250000
+      },
+      {
+        "GrLivArea": 1362,
+        "SalePrice": 143000
+      },
+      {
+        "GrLivArea": 1694,
+        "SalePrice": 307000
+      },
+      {
+        "GrLivArea": 2090,
+        "SalePrice": 200000
+      }
+    ]
+  }
